use event to initialize selected records on datwhen data loaded
