+++ lib/FindLocation/useLocationsRecords.js
@@ -1,7 +1,11 @@
 import omit from 'lodash/omit';
 import { useEffect, useState } from 'react';
 
-import { useLocations } from '../hooks';
+import { EVENT_EMMITER_EVENTS } from '../constants';
+import {
+  useEventEmitter,
+  useLocations,
+} from '../hooks';
 import { FILTERS_CONFIG } from './configs';
 import { filterAndSort } from './filterAndSort';
 
@@ -25,16 +29,36 @@ const dehydrateLocations = (hydratedLocations) => {
   return hydratedLocations.map((location) => omit(location, ['isAssigned']));
 };
 
+const emitInitialSelectedRecords = (eventEmitter, initialSelected, records) => {
+  if (!initialSelected?.length) return;
+
+  const recordsMap = initialSelected.reduce((acc, curr) => {
+    const record = records.find(({ id }) => id === curr);
+
+    if (record) {
+      acc[curr] = record;
+    }
+
+    return acc;
+  }, {});
+
+  return eventEmitter.emit(EVENT_EMMITER_EVENTS.INITIALIZE_SELECTED_RECORDS_MAP, recordsMap);
+}
+
 export const useLocationsRecords = ({
   filters,
   selectedLocations,
+  initialSelected,
   institutionsMap,
   campusessMap,
   librariesMap,
 }) => {
+  const eventEmitter = useEventEmitter();
   const [isProcessing, setIsProcessing] = useState(false);
   const [locationRecords, setLocationRecords] = useState([]);
-  const { locations, isLoading: isLocationsLoading } = useLocations();
+  const { locations, isLoading: isLocationsLoading } = useLocations({
+    onSuccess: emitInitialSelectedRecords.bind(null, eventEmitter, initialSelected),
+  });
 
   useEffect(() => {
     Promise.resolve(true)
