+++ lib/FindRecords/hooks/useRecordsSelect/useRecordsSelect.js
@@ -30,6 +30,30 @@ export const useRecordsSelect = ({ records }) => {
   const [selectedRecordsMap, setSelectedRecordsMap] = useState({});
   const selectedRecordsLength = Object.keys(selectedRecordsMap).length;
 
+  const emitChangeEvent = useCallback((recordsMap) => {
+    eventEmitter.emit(
+      EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED,
+      recordsMap,
+    );
+  }, [eventEmitter]);
+
+  /*
+    Subscribe to event for initialize selected records.
+  */
+  useEffect(() => {
+    const eventType = EVENT_EMMITER_EVENTS.INITIALIZE_SELECTED_RECORDS_MAP;
+    const callback = ({ detail: recordsMap }) => {
+      emitChangeEvent(recordsMap);
+      setSelectedRecordsMap(recordsMap);
+    };
+
+    eventEmitter.on(eventType, callback);
+
+    return () => {
+      eventEmitter.off(eventType, callback);
+    }
+  }, [emitChangeEvent]);
+
   useEffect(() => {
     setAllRecordsSelected(records.every(r => Boolean(selectedRecordsMap[r.id])));
   }, [records, selectedRecordsMap]);
@@ -39,16 +63,13 @@ export const useRecordsSelect = ({ records }) => {
       const newMap = {
         ...selectedFromOtherPages(selectedRecordsMap, records),
         ...reduceCheckedRecords(records, !allRecordsSelected),
-      }
+      };
 
-      eventEmitter.emit(
-        EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED,
-        newMap,
-      );
+      emitChangeEvent(newMap);
 
       return newMap;
     });
-  }, [allRecordsSelected, records, selectedRecordsMap]);
+  }, [allRecordsSelected, emitChangeEvent, records, selectedRecordsMap]);
 
   const selectRecord = useCallback(record => {
     const { id } = record;
@@ -63,14 +84,11 @@ export const useRecordsSelect = ({ records }) => {
         newMap[id] = record;
       }
 
-      eventEmitter.emit(
-        EVENT_EMMITER_EVENTS.FIND_RECORDS_SELECTED_CHANGED,
-        newMap,
-      );
+      emitChangeEvent(newMap);
 
       return newMap;
     });
-  }, []);
+  }, [emitChangeEvent]);
 
   const isRecordSelected = useCallback(
     ({ item }) => Boolean(selectedRecordsMap[item.id]),
