+++ lib/constants/events.js
@@ -1,3 +1,4 @@
 export const EVENT_EMMITER_EVENTS = {
   FIND_RECORDS_SELECTED_CHANGED: 'FIND_RECORDS_SELECTED_CHANGED',
+  INITIALIZE_SELECTED_RECORDS_MAP: 'INITIALIZE_SELECTED_RECORDS_MAP',
 };
