+++ lib/hooks/useLocations/useLocations.js
@@ -11,7 +11,7 @@ import { fetchAllRecords } from '../../utils';
 
 const DEFAULT_DATA = [];
 
-export const useLocations = () => {
+export const useLocations = (options) => {
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: 'locations' });
 
@@ -32,7 +32,8 @@ export const useLocations = () => {
         },
         'cql.allRecords=1 sortby name',
       )
-    }
+    },
+    options,
   );
 
   return ({
